+++ src/Commands/MakeSwarmBlueprintCommand.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace BuiltByBerry\LaravelSwarm\Commands;
 
+use BuiltByBerry\LaravelSwarm\Commands\Concerns\DetectsInteractiveConsole;
 use BuiltByBerry\LaravelSwarm\Commands\Concerns\InteractsWithBlueprintCorpus;
 use BuiltByBerry\LaravelSwarm\Commands\Concerns\ResolvesHostRootNamespace;
 use Illuminate\Console\Command;
@@ -43,6 +44,7 @@ use function Laravel\Prompts\text;
 #[AsCommand(name: 'make:swarm:blueprint')]
 class MakeSwarmBlueprintCommand extends Command
 {
+    use DetectsInteractiveConsole;
     use InteractsWithBlueprintCorpus;
     use ResolvesHostRootNamespace;
 
@@ -334,7 +336,7 @@ class MakeSwarmBlueprintCommand extends Command
      */
     private function wantsInteraction(): bool
     {
-        return $this->input->isInteractive() && $this->option('no-interaction') !== true;
+        return $this->consoleCanPrompt() && $this->option('no-interaction') !== true;
     }
 
     /**